chore: move tests & rename package
