+++ src/test/java/gift/integrationTest/ProductDomainTest.java
@@ -1,4 +1,4 @@
-package gift.domain.controller;
+package gift.integrationTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -28,7 +28,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @Transactional
-class ProductControllerTest {
+class ProductDomainTest {
 
     @LocalServerPort
     private int port;
